+++ CHANGES.md
@@ -88,6 +88,10 @@ To be released.
     Predicate<Currency>)` to `BlockChain<T>.ProposeGenesisBlock(
     PrivateKey, ImmutableList<Transaction<T>>, DateTimeOffset?, IAction,
     Predicate<Currency>)`.  [[#3079]]
+ -  (Libplanet.Explorer) `Currency.decimalPlaces`' type became `Byte`
+    from `UInt`.  [[#3085]]
+ -  (Libplanet.Explorer) `currencyHash` parameter was removed from
+    `balance` and `totalSupply` in `StateQuery` type.  [[#3085]]
 
 ### Backward-incompatible network protocol changes
 
@@ -150,6 +154,9 @@ To be released.
     interface.  [[#3084]]
  -  (Libplanet.Tools) Added `-v`/`--validator-key` option to
     `planet block generate-genesis` command.  [[#3088]]
+ -  (Libplanet.Explorer) Added `CurrencyInput` type.  [[#3085]]
+ -  (Libplanet.Explorer) Added `CurrencyInput`-typed `currency` parameter to
+    `balance` and `totalSupply` in `StateQuery` type.  [[#3085]]
 
 ### Behavioral changes
 
@@ -267,6 +274,7 @@ To be released.
 [#3079]: https://github.com/planetarium/libplanet/pull/3079
 [#3080]: https://github.com/planetarium/libplanet/pull/3080
 [#3084]: https://github.com/planetarium/libplanet/pull/3084
+[#3085]: https://github.com/planetarium/libplanet/pull/3085
 [#3088]: https://github.com/planetarium/libplanet/pull/3088